requirements.txt
matplotlib~=3.5.1
tqdm~=4.64.0
requests~=2.27.1
tensorflow~=2.8.0
Pillow~=9.1.0
beautifulsoup4~=4.11.1